+++ packages/content/src/lib/content.spec.ts
@@ -24,7 +24,7 @@ describe('injectContent', () => {
     injectContent().subscribe((c) => {
       expect(c.content).toMatch('No Content Found');
       expect(c.attributes).toEqual({});
-      expect(c.filename).toEqual('/src/content/test.md');
+      expect(c.filename).toEqual('/src/content/test');
     });
     flushMicrotasks();
     flush();
@@ -41,7 +41,7 @@ describe('injectContent', () => {
     injectContent().subscribe((c) => {
       expect(c.content).toMatch(customFallback);
       expect(c.attributes).toEqual({});
-      expect(c.filename).toEqual('/src/content/test.md');
+      expect(c.filename).toEqual('/src/content/test');
     });
     flushMicrotasks();
     flush();
@@ -68,7 +68,7 @@ Test Content`),
     injectContent().subscribe((c) => {
       expect(c.content).toMatch('Test Content');
       expect(c.attributes).toEqual({ slug: 'test' });
-      expect(c.filename).toEqual('/src/content/test.md');
+      expect(c.filename).toEqual('/src/content/test');
       expect(c.slug).toEqual('test');
     });
     flushMicrotasks();
@@ -98,7 +98,7 @@ Test Content`),
     injectContent().subscribe((c) => {
       expect(c.content).toMatch('Test Content');
       expect(c.attributes).toEqual({ slug: 'custom-slug-test' });
-      expect(c.filename).toEqual('/src/content/custom-slug-test.md');
+      expect(c.filename).toEqual('/src/content/custom-slug-test');
       expect(c.slug).toEqual('custom-slug-test');
     });
     flushMicrotasks();
@@ -129,7 +129,7 @@ Test Content`),
       expect(c.content).toMatch('Test Content');
       expect(c.attributes).toEqual({ slug: 'custom-prefix-slug-test' });
       expect(c.filename).toEqual(
-        '/src/content/customPrefix/custom-prefix-slug-test.md'
+        '/src/content/customPrefix/custom-prefix-slug-test'
       );
       expect(c.slug).toEqual('custom-prefix-slug-test');
     });
@@ -160,8 +160,61 @@ Test Content`),
     injectContent().subscribe((c) => {
       expect(c.content).toMatch('Test Content');
       expect(c.attributes).toEqual({ slug: 'custom-filename-test-slug' });
-      expect(c.filename).toEqual('/src/content/custom-filename-test.md');
+      expect(c.filename).toEqual('/src/content/custom-filename-test');
       expect(c.slug).toEqual('custom-filename-test');
+    });
+    flushMicrotasks();
+    flush();
+  }));
+
+  it('should return ContentFile object matching md file if both md and agx available', fakeAsync(() => {
+    const routeParams = { slug: 'test' };
+    const contentFiles = {
+      '/src/content/test.md': () =>
+        Promise.resolve(`---
+slug: 'test'
+---
+Test md Content`),
+      '/src/content/test.agx': () =>
+        Promise.resolve(`---
+slug: 'test'
+---
+Test agx Content`),
+    };
+    const { injectContent } = setup({
+      routeParams,
+      contentFiles,
+    });
+    injectContent().subscribe((c) => {
+      expect(c.content).toMatch('Test md Content');
+      expect(c.attributes).toEqual({ slug: 'test' });
+      expect(c.filename).toEqual('/src/content/test');
+      expect(c.slug).toEqual('test');
+    });
+    flushMicrotasks();
+    flush();
+  }));
+
+  it('should return ContentFile object with data set to exports for non-string contentFiles', fakeAsync(() => {
+    const routeParams = { slug: 'test' };
+    const metadataExport = { title: 'test' };
+    const defaultExport = 'default';
+    const contentFiles = {
+      '/src/content/test.agx': () =>
+        Promise.resolve({
+          default: defaultExport,
+          metadata: metadataExport,
+        }),
+    };
+    const { injectContent } = setup({
+      routeParams,
+      contentFiles,
+    });
+    injectContent().subscribe((c) => {
+      expect(c.content).toMatch(defaultExport);
+      expect(c.attributes).toEqual(metadataExport);
+      expect(c.filename).toEqual('/src/content/test');
+      expect(c.slug).toEqual('test');
     });
     flushMicrotasks();
     flush();
@@ -175,7 +228,10 @@ Test Content`),
         | { customFilename: string };
       customFallback: string;
       routeParams: { [key: string]: any };
-      contentFiles: Record<string, () => Promise<string>>;
+      contentFiles: Record<
+        string,
+        () => Promise<string | { default: any; metadata: any }>
+      >;
     }>
   ) {
     TestBed.configureTestingModule({