+++ packages/vite-plugin-nitro/src/lib/vite-plugin-nitro.ts
@@ -75,7 +75,12 @@ export function nitro(options?: Options, nitroOptions?: NitroConfig): Plugin[] {
         const rendererEntry =
           filePrefix +
           normalizePath(
-            join(__dirname, `runtime/renderer${filePrefix ? '.mjs' : ''}`)
+            join(
+              __dirname,
+              `runtime/renderer${!options?.ssr ? '-client' : ''}${
+                filePrefix ? '.mjs' : ''
+              }`
+            )
           );
 
         nitroConfig = {
@@ -124,6 +129,10 @@ export function nitro(options?: Options, nitroOptions?: NitroConfig): Plugin[] {
           nitroConfig = withVercelOutputAPI(nitroConfig, workspaceRoot);
         }
 
+        if (isCloudflarePreset(buildPreset)) {
+          nitroConfig = withCloudflareOutput(nitroConfig);
+        }
+
         if (!ssrBuild && !isTest) {
           // store the client output path for the SSR build config
           clientOutputPath = resolve(
@@ -144,6 +153,15 @@ export function nitro(options?: Options, nitroOptions?: NitroConfig): Plugin[] {
         };
 
         if (isBuild) {
+          nitroConfig.publicAssets = [{ dir: clientOutputPath }];
+          nitroConfig.serverAssets = [
+            {
+              baseName: 'public',
+              dir: clientOutputPath,
+            },
+          ];
+          nitroConfig.renderer = rendererEntry;
+
           if (isEmptyPrerenderRoutes(options)) {
             nitroConfig.prerender = {};
             nitroConfig.prerender.routes = ['/'];
@@ -196,13 +214,6 @@ export function nitro(options?: Options, nitroOptions?: NitroConfig): Plugin[] {
           if (ssrBuild) {
             nitroConfig = {
               ...nitroConfig,
-              publicAssets: [{ dir: clientOutputPath }],
-              serverAssets: [
-                {
-                  baseName: 'public',
-                  dir: clientOutputPath,
-                },
-              ],
               externals: {
                 ...nitroOptions?.externals,
                 external: [
@@ -216,7 +227,6 @@ export function nitro(options?: Options, nitroOptions?: NitroConfig): Plugin[] {
                 'zone.js/fesm2015/zone-node',
                 ...(nitroOptions?.moduleSideEffects || []),
               ],
-              renderer: rendererEntry,
               handlers: [
                 {
                   handler: apiMiddlewareHandler,
@@ -321,5 +331,17 @@ const withVercelOutputAPI = (
     publicDir: normalizePath(
       resolve(workspaceRoot, '.vercel', 'output/static')
     ),
+  },
+});
+
+const isCloudflarePreset = (buildPreset: string | undefined) =>
+  process.env['CF_PAGES'] ||
+  (buildPreset && buildPreset.toLowerCase().includes('cloudflare-pages'));
+
+const withCloudflareOutput = (nitroConfig: NitroConfig | undefined) => ({
+  ...nitroConfig,
+  output: {
+    ...nitroConfig?.output,
+    serverDir: '{{ output.publicDir }}/_worker.js',
   },
 });