+++ tools/publish.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 echo "Publishing to $RELEASE_TAG"
-#npm config set //registry.npmjs.org/:_authToken=$NPM_TOKEN
+npm config set //registry.npmjs.org/:_authToken=$NPM_TOKEN
 npm publish node_modules/@analogjs/astro-angular --access public --tag $RELEASE_TAG
 npm publish node_modules/@analogjs/content --access public --tag $RELEASE_TAG
 npm publish node_modules/@analogjs/platform --access public --tag $RELEASE_TAG